aesthetic fixes

+++ test/test_tools_ezmatio_1.cxx
@@ -46,7 +46,7 @@ BOOST_AUTO_TEST_CASE(basic)
   typedef Tomographer::MAT::DimList DimList;
   DimList dims{{3, 4, 5}};
   const std::vector<int> ok{{3, 4, 5}};
-  BOOST_CHECK_EQUAL(dims.size(), 3);
+  BOOST_CHECK_EQUAL(dims.size(), 3u);
   BOOST_CHECK_EQUAL(dims.ndims(), 3);
   BOOST_CHECK_EQUAL(dims.numel(), 3*4*5);
   BOOST_CHECK(dims == ok);
@@ -160,7 +160,7 @@ BOOST_AUTO_TEST_CASE(IndexListIterator1)
 
   for (std::size_t j = 0; j < (std::size_t)it.numel(); ++j) {
     BOOST_CHECK(it.valid());
-    BOOST_CHECK_EQUAL(it.linearIndex(), j);
+    BOOST_CHECK_EQUAL((long)it.linearIndex(), (long)j);
 
     BOOST_MESSAGE("j = " << j << ", index=" << it) ;
 
@@ -190,7 +190,7 @@ BOOST_AUTO_TEST_CASE(IndexListIterator2)
 
   for (std::size_t j = 0; j < (std::size_t)it.numel(); ++j) {
     BOOST_CHECK(it.valid());
-    BOOST_CHECK_EQUAL(it.linearIndex(), j);
+    BOOST_CHECK_EQUAL((long)it.linearIndex(), (long)j);
 
     std::vector<int> indlist(dims.size());
     std::size_t jj = j;
